Refactored InstanceProvider
