+++ collect_app/src/main/java/org/odk/collect/android/provider/InstanceProvider.java
@@ -21,14 +21,14 @@ import android.content.UriMatcher;
 import android.database.Cursor;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteException;
 import android.database.sqlite.SQLiteQueryBuilder;
 import android.net.Uri;
+import android.support.annotation.NonNull;
 import android.text.TextUtils;
 
 import org.odk.collect.android.R;
 import org.odk.collect.android.application.Collect;
-import org.odk.collect.android.database.ODKSQLiteOpenHelper;
+import org.odk.collect.android.database.helpers.InstancesDatabaseHelper;
 import org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns;
 import org.odk.collect.android.utilities.MediaUtils;
 
@@ -40,16 +40,10 @@ import java.util.Locale;
 
 import timber.log.Timber;
 
-/**
- *
- */
+import static org.odk.collect.android.database.helpers.InstancesDatabaseHelper.DATABASE_NAME;
+import static org.odk.collect.android.database.helpers.InstancesDatabaseHelper.INSTANCES_TABLE_NAME;
+
 public class InstanceProvider extends ContentProvider {
-
-
-    private static final String DATABASE_NAME = "instances.db";
-    private static final int DATABASE_VERSION = 4;
-    private static final String INSTANCES_TABLE_NAME = "instances";
-
     private static HashMap<String, String> sInstancesProjectionMap;
 
     private static final int INSTANCES = 1;
@@ -57,108 +51,9 @@ public class InstanceProvider extends ContentProvider {
 
     private static final UriMatcher sUriMatcher;
 
-    /**
-     * This class helps open, create, and upgrade the database file.
-     */
-    private static class DatabaseHelper extends ODKSQLiteOpenHelper {
+    private InstancesDatabaseHelper databaseHelper;
 
-        DatabaseHelper(String databaseName) {
-            super(Collect.METADATA_PATH, databaseName, null, DATABASE_VERSION);
-        }
-
-        @Override
-        public void onCreate(SQLiteDatabase db) {
-            db.execSQL("CREATE TABLE " + INSTANCES_TABLE_NAME + " ("
-                    + InstanceColumns._ID + " integer primary key, "
-                    + InstanceColumns.DISPLAY_NAME + " text not null, "
-                    + InstanceColumns.SUBMISSION_URI + " text, "
-                    + InstanceColumns.CAN_EDIT_WHEN_COMPLETE + " text, "
-                    + InstanceColumns.INSTANCE_FILE_PATH + " text not null, "
-                    + InstanceColumns.JR_FORM_ID + " text not null, "
-                    + InstanceColumns.JR_VERSION + " text, "
-                    + InstanceColumns.STATUS + " text not null, "
-                    + InstanceColumns.LAST_STATUS_CHANGE_DATE + " date not null, "
-                    + InstanceColumns.DISPLAY_SUBTEXT + " text not null,"
-                    + InstanceColumns.DELETED_DATE + " date );");
-        }
-
-        @SuppressWarnings({"checkstyle:FallThrough"})
-        @Override
-        public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-            Timber.i("Upgrading database from version %d to %d", oldVersion, newVersion);
-
-            boolean success = true;
-            switch (oldVersion) {
-                case 1:
-                    success = upgradeToVersion2(db);
-                case 2:
-                    success &= upgradeToVersion3(db);
-                case 3:
-                    success &= upgradeToVersion4(db);
-                    break;
-                default:
-                    Timber.i("Unknown version " + oldVersion);
-            }
-
-            if (success) {
-                Timber.i("Upgrading database from version " + oldVersion + " to " + newVersion + " completed with success.");
-            } else {
-                Timber.i("Upgrading database from version " + oldVersion + " to " + newVersion + " failed.");
-            }
-        }
-
-        private boolean upgradeToVersion2(SQLiteDatabase db) {
-            boolean success = true;
-            try {
-                db.execSQL("ALTER TABLE " + INSTANCES_TABLE_NAME + " ADD COLUMN "
-                        + InstanceProviderAPI.InstanceColumns.CAN_EDIT_WHEN_COMPLETE + " text;");
-                db.execSQL("UPDATE " + INSTANCES_TABLE_NAME + " SET "
-                        + InstanceProviderAPI.InstanceColumns.CAN_EDIT_WHEN_COMPLETE + " = '" + Boolean.toString(true)
-                        + "' WHERE " + InstanceProviderAPI.InstanceColumns.STATUS + " IS NOT NULL AND "
-                        + InstanceProviderAPI.InstanceColumns.STATUS + " != '" + InstanceProviderAPI.STATUS_INCOMPLETE
-                        + "'");
-            } catch (SQLiteException e) {
-                Timber.e(e);
-                success = false;
-            }
-            return success;
-        }
-
-        private boolean upgradeToVersion3(SQLiteDatabase db) {
-            boolean success = true;
-            try {
-                db.execSQL("ALTER TABLE " + INSTANCES_TABLE_NAME + " ADD COLUMN "
-                        + InstanceProviderAPI.InstanceColumns.JR_VERSION + " text;");
-            } catch (SQLiteException e) {
-                Timber.e(e);
-                success = false;
-            }
-            return success;
-        }
-
-        private boolean upgradeToVersion4(SQLiteDatabase db) {
-            boolean success = true;
-            try {
-                Cursor cursor = db.rawQuery("SELECT * FROM " + INSTANCES_TABLE_NAME + " LIMIT 0", null);
-                int columnIndex = cursor.getColumnIndex(InstanceProviderAPI.InstanceColumns.DELETED_DATE);
-                cursor.close();
-
-                // Only add the column if it doesn't already exist
-                if (columnIndex == -1) {
-                    db.execSQL("ALTER TABLE " + INSTANCES_TABLE_NAME + " ADD COLUMN "
-                            + InstanceProviderAPI.InstanceColumns.DELETED_DATE + " date;");
-                }
-            } catch (SQLiteException e) {
-                Timber.e(e);
-                success = false;
-            }
-            return success;
-        }
-    }
-
-    private DatabaseHelper databaseHelper;
-
-    private DatabaseHelper getDbHelper() {
+    private InstancesDatabaseHelper getDbHelper() {
         // wrapper to test and reset/set the dbHelper based upon the attachment state of the device.
         try {
             Collect.createODKDirs();
@@ -170,20 +65,19 @@ public class InstanceProvider extends ContentProvider {
         if (databaseHelper != null) {
             return databaseHelper;
         }
-        databaseHelper = new DatabaseHelper(DATABASE_NAME);
+        databaseHelper = new InstancesDatabaseHelper(DATABASE_NAME);
         return databaseHelper;
     }
 
     @Override
     public boolean onCreate() {
         // must be at the beginning of any activity that can be called from an external intent
-        DatabaseHelper h = getDbHelper();
+        InstancesDatabaseHelper h = getDbHelper();
         return h != null;
     }
 
-
     @Override
-    public Cursor query(Uri uri, String[] projection, String selection, String[] selectionArgs,
+    public Cursor query(@NonNull Uri uri, String[] projection, String selection, String[] selectionArgs,
                         String sortOrder) {
         SQLiteQueryBuilder qb = new SQLiteQueryBuilder();
         qb.setTables(INSTANCES_TABLE_NAME);
@@ -211,9 +105,8 @@ public class InstanceProvider extends ContentProvider {
         return c;
     }
 
-
     @Override
-    public String getType(Uri uri) {
+    public String getType(@NonNull Uri uri) {
         switch (sUriMatcher.match(uri)) {
             case INSTANCES:
                 return InstanceColumns.CONTENT_TYPE;
@@ -226,9 +119,8 @@ public class InstanceProvider extends ContentProvider {
         }
     }
 
-
     @Override
-    public Uri insert(Uri uri, ContentValues initialValues) {
+    public Uri insert(@NonNull Uri uri, ContentValues initialValues) {
         // Validate the requested uri
         if (sUriMatcher.match(uri) != INSTANCES) {
             throw new IllegalArgumentException("Unknown URI " + uri);
@@ -321,14 +213,13 @@ public class InstanceProvider extends ContentProvider {
         }
     }
 
-
     /**
      * This method removes the entry from the content provider, and also removes any associated
      * files.
      * files:  form.xml, [formmd5].formdef, formname-media {directory}
      */
     @Override
-    public int delete(Uri uri, String where, String[] whereArgs) {
+    public int delete(@NonNull Uri uri, String where, String[] whereArgs) {
         SQLiteDatabase db = getDbHelper().getWritableDatabase();
         int count;
 
@@ -337,7 +228,7 @@ public class InstanceProvider extends ContentProvider {
                 Cursor del = null;
                 try {
                     del = this.query(uri, null, where, whereArgs, null);
-                    if (del.getCount() > 0) {
+                    if (del != null && del.getCount() > 0) {
                         del.moveToFirst();
                         do {
                             String instanceFile = del.getString(
@@ -363,7 +254,7 @@ public class InstanceProvider extends ContentProvider {
                 String status = null;
                 try {
                     c = this.query(uri, null, where, whereArgs, null);
-                    if (c.getCount() > 0) {
+                    if (c != null && c.getCount() > 0) {
                         c.moveToFirst();
                         status = c.getString(c.getColumnIndex(InstanceColumns.STATUS));
                         do {
@@ -404,9 +295,8 @@ public class InstanceProvider extends ContentProvider {
         return count;
     }
 
-
     @Override
-    public int update(Uri uri, ContentValues values, String where, String[] whereArgs) {
+    public int update(@NonNull Uri uri, ContentValues values, String where, String[] whereArgs) {
         SQLiteDatabase db = getDbHelper().getWritableDatabase();
 
         Long now = System.currentTimeMillis();
@@ -417,7 +307,7 @@ public class InstanceProvider extends ContentProvider {
         }
 
         int count;
-        String status = null;
+        String status;
         switch (sUriMatcher.match(uri)) {
             case INSTANCES:
                 if (values.containsKey(InstanceColumns.STATUS)) {
@@ -466,7 +356,7 @@ public class InstanceProvider extends ContentProvider {
         sUriMatcher.addURI(InstanceProviderAPI.AUTHORITY, "instances", INSTANCES);
         sUriMatcher.addURI(InstanceProviderAPI.AUTHORITY, "instances/#", INSTANCE_ID);
 
-        sInstancesProjectionMap = new HashMap<String, String>();
+        sInstancesProjectionMap = new HashMap<>();
         sInstancesProjectionMap.put(InstanceColumns._ID, InstanceColumns._ID);
         sInstancesProjectionMap.put(InstanceColumns.DISPLAY_NAME, InstanceColumns.DISPLAY_NAME);
         sInstancesProjectionMap.put(InstanceColumns.SUBMISSION_URI, InstanceColumns.SUBMISSION_URI);
@@ -483,5 +373,4 @@ public class InstanceProvider extends ContentProvider {
                 InstanceColumns.DISPLAY_SUBTEXT);
         sInstancesProjectionMap.put(InstanceColumns.DELETED_DATE, InstanceColumns.DELETED_DATE);
     }
-
 }
